requirements.txt
numpy==1.21.5
scipy==1.7.3
scikit-learn==1.0.2
joblib==1.1.0
librosa==0.9.1
torchaudio==0.8.0